requirements.txt
argon2-cffi==18.1.0
bcrypt==3.1.4
certifi==2017.11.5
cffi==1.11.4
Django==2.0.2
django-autocomplete-light==3.2.10
django-bootstrap3==9.1.0
django-braces==1.12.0
django-widget-tweaks==1.4.1
Faker==0.8.10
djangorestframework==3.8.2
Pillow==5.0.0
psycopg2==2.7.3.2
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
text-unidecode==1.1
wincertstore==0.2
python-dotenv
django-easy-pdf==0.2.0
WeasyPrint>=0.34